requirements.txt
tensorflow==2.10.1
tensorflow-gpu==2.10.1
pandas
Pillow
scikit-learn
matplotlib
tensorboard
opencv-python